Add Wake Up and Sleep Times
This PR is based off of @johneagle's PR (https://github.com/hubot-scripts/hubot-heroku-keepalive/pull/9). It takes the code he had written, and modifies it slightly to not make HUBOT_HEROKU_WAKEUP_TIME and HUBOT_HEROKU_SLEEP_TIME truly optional. If both of those parameters are not in place, then it works as it previously did instead of setting defaults for those items.